requirements.txt
spotipy>=2.23.0
lyricsgenius>=3.0.1
scikit-learn>=1.4.0
pandas>=2.2.0
nltk>=3.8.1
python-dotenv>=1.0.1
streamlit>=1.33.0
